requirements.txt
qdrant-client
streamlit
fastembed
python-dotenv
newsapi-python
